rexport macro
